test(optimum): fold Qwen3.5 param-guard tests into test_params.py
Move TestGetParamQwen35 out of the separate test_multimodal_params.py into the
existing converter/test_params.py (both are optimum compile-param unit tests)
and delete the now-redundant file.
